Scraping script adjustment, preparation for ROS
I adjusted the script so that it obtains more accurate dates, as the
landing pages of some of the links on the main UTCS events page weren't
updated and thus were giving incorrect dates. I did this by scraping
dates from the main page, where they are correct and up-to-date.
In other news, I think that integrating the Express Server with ROS
should be doable. I'm going to head to the lab today to see how much
headway I can make with it.