Clicking a button now publishes room# to topic
Specifically, this version (minus the uncomments in app.js regarding
ROS, of course) should now publish a topic, /roomNumber, and whenever a
button is pressed on the HTML page, it makes a POST request, which
in turn triggers the publishing of a message (the room number) by the
created rostopic. Hurray!